requirements.txt
pandas==1.3.5
numpy==1.21.5
scikit-learn==1.0.2
xgboost==1.5.2
lightgbm==3.3.2
tensorflow==2.8.0
optuna==2.10.0
joblib==1.1.0
matplotlib==3.5.1
pyyaml==6.0
dask==2022.2.1
numba==0.55.1
